reworked HomeworkList structure
